✨ handle successful authentication
Code is almost copied/pasted from the OpenId Passport Strategy.
We handle successful authentication by running a get or create function,
that looks up for the user by their email, or create a new user if not
found.

Agent Connect recommends to rely on the email and not the sub.

Please note, some works need to be done to arrange a proper migration.
The current production environment use MCP as OIDC provider. If they
identify the user based on its sub, we should need to prepare the production
data to be able to reconciliate user based on their email.

Passport is responsible of managing user session.