♻️ manage manually session data
When adding the logout feature, we will need to save some data in the
session while logging the user, and access it while loging out the user.

By default, when Passport authenticate is called, it resets the session
data.

Prevent this default behavior, and take care to delete any additional session
data added while logging the user.