✨ persist id_token when logging the user in their session data
This token will be required when logging out the user from the OIDC
provider. Agent connect required an id_token_hint.

The id_token is only available when logging the user, thus we need to
store its value in the session store, to make sure, when the user
call the logout route, which is another route, we can resolve its
id token.