✨logout the user from Agent Connect
Based on Agent Connect requirements, logout the user from the OIDC provider.
The logout flow uses to step.

We initiate the logout by redirecting the user agent to the OIDC provider.
The OIDC provider validate our request, and clear its session cookie for the
given user (session cookies are set for the OIDC provider domain), then the
OP redirect the user back to our backend, we handle the callback, and end
its session in our app.